python_api/requirements.txt
flask
torch
transformers
pillow